refactor: move functions back to invitations but outside of components
